# tratamientos/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/treatments/application": {
            "put": {
                "description": "Updates an application changing everything on it, except the id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Updates an application",
                "parameters": [
                    {
                        "description": "TBD",
                        "name": "Application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an application for a given animal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Creates an application",
                "parameters": [
                    {
                        "description": "TBD",
                        "name": "Application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/application/pet/{pet}": {
            "get": {
                "description": "Updates an application only changing the specified fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Get an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application affected",
                        "name": "pet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.Application"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/application/specific/{id}": {
            "get": {
                "description": "Get an application with a given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Get an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/application/treatment/{treatmentId}": {
            "get": {
                "description": "Updates an application only changing the specified fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Get an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application affected",
                        "name": "treatmentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.Application"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/application/{id}": {
            "delete": {
                "description": "Removes a given application with given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Deletes an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application affected",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates an application only changing the specified fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application request"
                ],
                "summary": "Updates an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application affected",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TBD",
                        "name": "Application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.Application": {
            "type": "object",
            "properties": {
                "applied_to": {
                    "type": "integer",
                    "example": 20
                },
                "date": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z"
                },
                "id": {
                    "type": "string",
                    "example": "e013f973-ed95-45c5-8bc4-3abf2d9112c3"
                },
                "name": {
                    "type": "string",
                    "example": "Anti rabica"
                },
                "treatment_id": {
                    "type": "string",
                    "example": "a45b9e1a-366a-450e-b298-f455139bfcd0"
                },
                "type": {
                    "type": "string",
                    "example": "vaccine"
                }
            }
        },
...
        "controller.ErrorMsg": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}